+++ tests/Unit/Service/EditorRendererTest.php
@@ -11,6 +11,7 @@ namespace OxidEsales\WysiwygModule\Tests\Unit\Service;
 
 use OxidEsales\Eshop\Core\ViewConfig;
 use OxidEsales\EshopCommunity\Internal\Framework\Templating\TemplateRendererInterface;
+use OxidEsales\WysiwygModule\HtmlFilter\HtmlFilterInterface;
 use OxidEsales\WysiwygModule\Service\EditorRenderer;
 use OxidEsales\WysiwygModule\Service\SettingsInterface;
 use PHPUnit\Framework\Attributes\DataProvider;
@@ -73,8 +74,15 @@ class EditorRendererTest extends TestCase
         string $expectedWidth,
         string $expectedHeight,
     ): void {
+
+        $templateRendererSpy = $this->createMock(TemplateRendererInterface::class);
+        $htmlFilterMock = $this->createMock(HtmlFilterInterface::class);
+        $htmlFilterMock
+            ->method('filter');
+
         $sut = $this->getSut(
-            templateRenderer: $templateRendererSpy = $this->createMock(TemplateRendererInterface::class)
+            templateRenderer: $templateRendererSpy,
+            htmlFilter: $htmlFilterMock
         );
 
         $templateRendererSpy
@@ -95,12 +103,14 @@ class EditorRendererTest extends TestCase
 
     public function testRenderCalledWithCorrectInputValues(): void
     {
-        $sut = $this->getSut(
-            templateRenderer: $templateRendererSpy = $this->createMock(TemplateRendererInterface::class)
-        );
+        $templateRendererSpy = $this->createMock(TemplateRendererInterface::class);
+        $htmlFilterStub = $this->createMock(HtmlFilterInterface::class);
 
         $fieldName = uniqid();
         $fieldValue = uniqid();
+
+        $htmlFilterStub->method('filter')
+            ->willReturn($fieldValue);
 
         $templateRendererSpy
             ->expects($this->once())
@@ -114,6 +124,11 @@ class EditorRendererTest extends TestCase
                     return true;
                 })
             );
+
+        $sut = $this->getSut(
+            templateRenderer: $templateRendererSpy,
+            htmlFilter: $htmlFilterStub
+        );
 
         $sut->render('any', 'any', $fieldValue, $fieldName);
     }
@@ -198,13 +213,67 @@ class EditorRendererTest extends TestCase
         $sut->render('any', 'any', 'any', 'any');
     }
 
-    public function getSut(
+    #[DataProvider('filterTemplateProvider')]
+    public function testFilterContent(string $template, string $expectedTemplate): void
+    {
+        $htmlFilterStub = $this->createMock(HtmlFilterInterface::class);
+        $htmlFilterStub
+            ->method('filter')
+            ->willReturn($expectedTemplate);
+
+        $templateRendererSpy = $this->createMock(TemplateRendererInterface::class);
+        $templateRendererSpy
+            ->expects($this->once())
+            ->method('renderTemplate')
+            ->with(
+                '@ddoewysiwyg/ddoewysiwyg',
+                $this->callback(function ($context) use ($expectedTemplate) {
+                    return $expectedTemplate == $context['sEditorValue'];
+                })
+            );
+
+        $sut = $this->getSut(
+            templateRenderer:  $templateRendererSpy,
+            htmlFilter: $htmlFilterStub
+        );
+        $sut->render('any', 'any', $template, 'any');
+    }
+
+    public static function filterTemplateProvider(): array
+    {
+        return [
+            [
+                'template' => 'plain template',
+                'expectedTemplate' => 'plain template',
+            ],
+            [
+                'template' => '<div>template</div>',
+                'expectedTemplate' => '<div>template</div>',
+            ],
+            [
+                'template' => '<p>par 1</p><script>//js1</script><p>par 2</p>',
+                'expectedTemplate' => '<p>par 1</p>//js1<p>par 2</p>',
+            ],
+            [
+                'template' => '<script>//js1</script><script>//js2</script>',
+                'expectedTemplate' => '//js1//js2',
+            ],
+            [
+                'template' => '<p>par 1</p><script src="app.js"/><p>par 2</p>',
+                'expectedTemplate' => '<p>par 1</p><p>par 2</p>',
+            ],
+        ];
+    }
+
+    private function getSut(
         TemplateRendererInterface $templateRenderer = null,
         SettingsInterface $settingsService = null,
+        HtmlFilterInterface $htmlFilter = null,
     ): EditorRenderer {
         return new EditorRenderer(
             templateRenderer: $templateRenderer ?? $this->createStub(TemplateRendererInterface::class),
             settingsService: $settingsService ?? $this->createStub(SettingsInterface::class),
+            htmlFilter: $htmlFilter ?? $this->createStub(HtmlFilterInterface::class),
         );
     }
 }